got tacs model working with caps wing examples
